+++ foundry/middleware.py
@@ -41,6 +41,11 @@ class AgeGateway:
         if re.match(PROTECTED_URLS_PATTERN, request.META['PATH_INFO']) is not None:
             return response
 
+        # Listing feeds also exempted
+        # todo: make the test more refined.
+        if request.META['PATH_INFO'].endswith('/feed/'):
+            return response
+
         # Now only do we hit the database
         # xxx: investigate preference caching. May want to hit the db less.
         private_site = preferences.GeneralPreferences.private_site
@@ -61,6 +66,17 @@ class AgeGateway:
                 re.match(
                     r'|'.join(exempted_urls.split()), 
                     request.META['PATH_INFO']
+               ) is not None
+            ):
+            return response
+
+        # Exempted IP addresses
+        exempted_ips = preferences.GeneralPreferences.exempted_ips
+        if exempted_ips \
+            and (
+                re.match(
+                    r'|'.join(exempted_ips.split()), 
+                    request.META['REMOTE_ADDR']
                ) is not None
             ):
             return response