+++ foundry/settings.py
@@ -11,8 +11,6 @@ import warnings
 
 
 FOUNDRY = {
-    'has_javascript': True,
-    'has_ajax': True,
     'sms_gateway_api_key': '',
     'sms_gateway_password': '',
     'layers': ('basic',)
@@ -202,7 +200,10 @@ COMMENTS_APP = 'foundry'
 SIMPLE_AUTOCOMPLETE = {
     'auth.user': {'threshold': 20},
     'category.category': {'threshold':20},
-    'jmbo.modelbase': {'threshold':50}
+    'jmbo.modelbase': {
+        'threshold': 50, 
+        'duplicate_format_function': lambda item, model, content_type: item.as_leaf_class().content_type.name
+    }
 }
 
 STATICFILES_FINDERS = (