+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 next
 ----
 #. Patch django.contrib.sites.models.Site.__unicode__ so it returns name and not domain. The UI gets confusing since we have up to three sites comprising one logical mobi site.
+#. Listings now have automatic RSS feeds.
 
 0.7.2
 -----