+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ next
 ----
 #. Patch django.contrib.sites.models.Site.__unicode__ so it returns name and not domain. The UI gets confusing since we have up to three sites comprising one logical mobi site.
 #. Listings now have automatic RSS feeds.
+#. Comment form now fires up correct virtual keyboard for a smart phone.
 
 0.7.2
 -----