+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ next
 #. Comment form now fires up correct virtual keyboard for a smart phone.
 #. Logged in members can now flag offensive comments. After three flags a moderator is notified.
 #. Some IP addresses can now be allowed to bypass the age gateway / private site. This allows the sharing of protected content via eg. Facebook and Twitter.
+#. Listing gets an optional RSS feed.
 
 0.7.2
 -----